+++ TensorIO/tensorio/src/main/java/ai/doc/tensorio/TIOUtilities/FileIO.java
@@ -1,4 +1,4 @@
-package ai.doc.tensorio.utils;
+package ai.doc.tensorio.TIOUtilities;
 
 import android.content.Context;
